Refactor vulnerability-slack-alert action
This update includes a new filter for the user to specify if they want to show the most recent alerts or most severe.
They can also decide how many alerts to show in each message sent to slack.
It now considers all code scanning alerts when constructing the message sent to slack to show the total amount of alerts and most important ones first. This was done to prevent developers from getting a false sense of security, as the previous version of this action didnt show alerts older than a specified date. This should give developers more clarification on alerts created.